+++ trex-lib/src/main/java/com/github/trex_paxos/Message.java
@@ -2,5 +2,5 @@ package com.github.trex_paxos;
 
 public interface Message {    
     /// @return the node in the cluster that sent this message.
-  byte from();
+    short from();
 }
