+++ trex-lib/src/main/java/com/github/trex_paxos/msg/AcceptResponse.java
@@ -24,16 +24,16 @@ package com.github.trex_paxos.msg;
 /// @param to                    see {@link DirectMessage}
 /// @param vote                  whether wre have voted for or voted against the Prepare message based on our past promises.
 /// @param highestFixedIndex additional information about the highest accepted index so that a leader will abdicate if it is behind.
-public record AcceptResponse(byte from,
-                             byte to,
+public record AcceptResponse(short from,
+                              short to,
                              Vote vote,
                              long highestFixedIndex
 ) implements TrexMessage, DirectMessage, LearningMessage {
   public record Vote(
       // spookily intellij says there are no usages of this field, but if I remove it everything breaks
-      byte from,
+      short from,
       // spookily intellij says there are no usages of this field, but if I remove it everything breaks
-      byte to,
+      short to,
       long logIndex,
       boolean vote
   ) {
