+++ trex-lib/src/test/java/com/github/trex_paxos/CatchupPropertyTests.java
@@ -32,12 +32,12 @@ public class CatchupPropertyTests {
 
   @Property(generation = GenerationMode.EXHAUSTIVE)
   void catchupTests(@ForAll("testCases") TestCase testCase) {
-    final var thisNodeId = (byte) 2;
+    final var thisNodeId = (short) 2;
 
     final var otherNodeId = switch (testCase.nodeIdentifierRelation) {
-      case LESS -> (byte) (thisNodeId - 1);
+      case LESS -> (short) (thisNodeId - 1);
       case EQUAL -> thisNodeId;
-      case GREATER -> (byte) (thisNodeId + 1);
+      case GREATER -> (short) (thisNodeId + 1);
     };
 
     final var thisCounter = 100;
