+++ trex-lib/src/test/java/com/github/trex_paxos/FakeJournal.java
@@ -22,7 +22,7 @@ public class FakeJournal implements Journal {
   }
 
   @Override
-  public Progress readProgress(byte nodeIdentifier) {
+  public Progress readProgress(short nodeIdentifier) {
     return new Progress(nodeIdentifier, promise, highestFixed);
   }
